balance will be agains TRY asset

+++ src/main/java/com/emce/brokage/balance/dto/BalanceRequest.java
@@ -1,8 +1,10 @@
 package com.emce.brokage.balance.dto;
 
-import com.emce.brokage.entity.TransactionType;
+import com.emce.brokage.balance.entity.TransactionType;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record BalanceRequest(Integer customerId, Double amount, TransactionType transactionType) {
+public record BalanceRequest(Integer customerId, BigDecimal amount, TransactionType transactionType) {
 }
