+++ src/main/java/com/emce/brokage/balance/dto/BalanceResponse.java
@@ -1,8 +1,10 @@
 package com.emce.brokage.balance.dto;
 
-import com.emce.brokage.entity.TransactionStatus;
+import com.emce.brokage.balance.entity.TransactionStatus;
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record BalanceResponse(Integer customerId, Double previousAmount, Double currentAmount, TransactionStatus transactionStatus) {
+public record BalanceResponse(Integer customerId, BigDecimal previousAmount, BigDecimal currentAmount, TransactionStatus transactionStatus) {
 }
