+++ src/main/java/com/emce/brokage/entity/Order.java
@@ -20,6 +20,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
@@ -38,8 +39,8 @@ public class Order {
     private String assetName;
     @Enumerated(EnumType.STRING)
     private OrderSide orderSide;
-    private Integer size;
-    private Double price;
+    private BigDecimal size;
+    private BigDecimal price;
     @Enumerated(EnumType.STRING)
     private OrderStatus status;
     @ManyToOne
